refactor: simplify logic
