+++ packages/apps/hr/src/components/templates/index.js
@@ -21,8 +21,8 @@ export {default as CityFormInput} from './CityFormInput/CityFormInput';
 export {default as CurrencySearchBar} from './CurrencySearchBar/CurrencySearchBar';
 export {default as DistanceIncrement} from './DistanceIncrement/DistanceIncrement';
 export {default as ExpenseHeader} from './ExpenseHeader/ExpenseHeader';
-export {default as ExpenseLineDisplayType} from './ExpenseLineDisplayType/ExpenseLineDisplayType';
 export {default as ExpenseLineSwitchAdd} from './ExpenseLineSwitchAdd/ExpenseLineSwitchAdd';
+export {default as ExpenseLineTypeDisplay} from './ExpenseLineTypeDisplay/ExpenseLineTypeDisplay';
 export {default as ExpenseLineTypeSwitch} from './ExpenseLineTypeSwitch/ExpenseLineTypeSwitch';
 export {default as ExpenseTypeSearchBar} from './ExpenseTypeSearchBar/ExpenseTypeSearchBar';
 export {default as ExpenseWaitingValidationSearchBar} from './ExpenseWaitingValidationSearchBar/ExpenseWaitingValidationSearchBar';
